Perform expression analysis on assumed-type dummies
When we recognized an expression as an assumed-type dummy argument
we did not otherwise analyze it, so `typedExpr` was not filled in.
When expression check is complete we expect every `typedExpr` data
member to have a value.

The fix is always to analyze the expression and then detect the
assumed-type dummy argument.

Fixes #915.