Add Fortran IR (FIR) MLIR dialect implementation
Adds FIR library that implements an MLIR dialect to which Fortran
parse-tree will be lowered to.
FIR is documented in FIRLangRef.md added with this commit.

Note:
This commit does not reflect an actual work log, it is a feature based split of the
changes done in the FIR experimental branch. The related work log can be found in the commits
between: https://github.com/schweitzpgi/f18/commit/8c320e3bf2c3e9cdac66c81db3bf4634bf972e1d
and:
https://github.com/schweitzpgi/f18/commit/9b9ea05f9a75608c7bb5372c56bf7b9363569a69