Add FIR optimizer library
This library provides transformation passes that can be run of FIR
to perform high level optimizations.
It aslo conatins a few utilities like a utility to mangle Fortran symbol
names inside FIR or another utility to map front-end type kinds to hardware types.

Note:
This commit does not reflect an actual work log, it is a feature based split of the
changes done in the FIR experimental branch. The related work log can be found in the commits between:
https://github.com/schweitzpgi/f18/commit/8c320e3bf2c3e9cdac66c81db3bf4634bf972e1d
and:
https://github.com/schweitzpgi/f18/commit/9b9ea05f9a75608c7bb5372c56bf7b9363569a69