Add bridge from Fortran parse-tree to FIR
- Add the `BurnsideBridge` class that implements the lowering of
the Fortran parse-tree to FIR operations as described in
BurnsidetoFIR.md.
- The lowering of IO statements is done in another file
(IO.cpp) to keep the bridge smaller.

Note:
This commit does not reflect an actual work log, it is a feature based split of the
changes done in the FIR experimental branch. The related work log can be found in the commits
between:
https://github.com/schweitzpgi/f18/commit/8c320e3bf2c3e9cdac66c81db3bf4634bf972e1d
and:
https://github.com/schweitzpgi/f18/commit/9b9ea05f9a75608c7bb5372c56bf7b9363569a69