Add bbc and tco tools
- Add bbc tool that takes a Fortran input file, drives parsing,
semantic analysis, and then lowers the parse-tree to FIR. It can
then run transformation passes on the obtained MLIR before
generating LLVM IR.

- Add tco tool that is meant to process mlir input files and
drives transformations on it.

Note:
This commit does not reflect an actual work log, it is a feature based split of the
changes done in the FIR experimental branch. The related work log can be found in the commits between:
https://github.com/schweitzpgi/f18/commit/8c320e3bf2c3e9cdac66c81db3bf4634bf972e1d
and:
https://github.com/schweitzpgi/f18/commit/9b9ea05f9a75608c7bb5372c56bf7b9363569a69