Add reference to the fortran frontend by setting it in the entrypoint. $ This patch sets flang frontend according to the name of the driver.
To test the entrypoint and keep the previous behavior of flang this patch adds
flang-tmp as a tool in lit configuration files and adds the binary location in
FLANG_TOOLS_DIR.

As consequence the tests were modified to look for flang-tmp. Moreover,
"--driver-mode=flang" was removed as the flag is sent by the entry point in
tools/flang/main.cpp.

Important:
The entry point uses the flag create by the PR https://reviews.llvm.org/D73951
to set fortran frontend.