Check for constrains C741 through C749
Most of these checks were already implemented and I just added references to
them to the code and tests.

I implemented the check for C747 to not allow coarray components in derived
types that are of type C_PTR, C_FUNPTR, or type TEAM_TYPE.

I implemented the check for C748 that requires a data component whose type has
a coarray ultimate component to be a nonpointer, nonallocatable scalar and not
be a coarray.

I also fixed some unrelated tests that got new errors when I implemented these
checks.