Responses to pull request comments
I redid the error messages for violations of C746 to be more consistent and
customized them so that we don't complain unnecessarily about deferred
coshapes.

I removed some spurious references to C746.

I fixed the error message for coarray components of type C_PTR or C_FUNPTR.

I added an attachment to the messages relating to ultimate components to point
to the relevant declarations.

I fixed the tests to match the message changes in the code and added another
test for components that are allocatable coarrays that do not have a deferred
coshape.